Add more scoring fields to Tournament object
Cutoff line is a percent measure of the number of scores to "drop"
from a team's ranked points. Subject to change: may become a flat cutoff
per tournament.

Cutoff score is the amount of team points to award players
who fall below the cutoff line for their team.

No show score is the amount of team points to award players
who do not post a score at all.