Sweeping changes largely stemming from implementing STI on Score -> PersonalBest, TournamentScore
Having two tables for two entities with all but a handful of
the same fields was stupid, so now TournamentScore
and PersonalBest extend a base 'Score' class.

They get to use the same table. They get to use the same
form class. It took a lot of tinkering with the association mapping
and small changes across the controllers/repos/templates but
I think it's going to be worth it.