Make editing Tournaments suck less
Step one is splitting "Games" away as its own form. Now the basic
details (start and end date, title) can be saved separately from
the collection of games, and updating the collection is much
less of a pain.

TODO: The same but with Teams. Tournament form now lists the teams,
will provide links to team_new and team_edit instead of trying
to manage it all in-place