Make managing teams less bad
Use a 'Roster' collection similar to the Games collection
on the tournament form. Adds a dialog that allows
a TO to either remove a player from a team,
or send the player to another team.

Required updates to associated routes. Also requierd a new
field on DraftEntry - eligible - as well as calls to set eligibility
when a User is added/removed from a team. In this
context, "eligible" means the player has a DraftEntry for
a given tournament, and that they have not already been
drafted (added to a team).