Big brain (?) architecture changes
TournamentUser extends from User, and serves as both a
reference to the User and a place to store all tournament
related data (Team, TournamentScores, stats like average rank, etc.)

Why? It simplifies a lot of the work of moving a "User" around a tournament
(moving from one team to another, signing up/withdrawing), makes
it faster to find stats about performance on a per tournament basis,
and most importantly facilitates running Tournaments that do not
require the User to be on a team.